refactor(node-ui): split PanelRight modules
