+++ packages/node-ui/src/ui/components/Shell/PanelRight/SessionsTab.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import { formatLocalTimestamp } from './format.js';
+import type { LocalAgentSessionSummary } from './types.js';
+
+export function SessionsTab(props: {
+  sessions: LocalAgentSessionSummary[];
+  onOpenSession: (session: LocalAgentSessionSummary) => void;
+}) {
+  const { sessions, onOpenSession } = props;
+
+  return (
+    <div className="v10-agent-scroll-tab">
+      <div className="v10-sessions-list">
+        <div className="v10-local-agent-copy" style={{ marginBottom: 12 }}>
+          Sessions track DKG-persisted conversations for your integrated agents.
+        </div>
+        {sessions.length === 0 ? (
+          <p className="v10-agent-empty-state">No integrated-agent sessions yet.</p>
+        ) : (
+          sessions.map((session) => (
+            <button
+              key={session.sessionId}
+              className="v10-session-item"
+              onClick={() => onOpenSession(session)}
+            >
+              <span className="v10-session-preview">
+                {session.integrationName}: {session.preview}
+              </span>
+              <span className="v10-session-count">
+                {session.messageCount} msg{session.messageCount !== 1 ? 's' : ''}
+                {session.lastTs ? ` - ${formatLocalTimestamp(session.lastTs)}` : ''}
+              </span>
+            </button>
+          ))
+        )}
+      </div>
+    </div>
+  );
+}
+
