+++ packages/node-ui/src/ui/components/Shell/PanelRight/constants.ts
@@ -1,0 +1,5 @@
+export const OPENCLAW_DOCS_URL = 'https://docs.openclaw.ai/';
+export const OPENCLAW_RELEASE_URL = 'https://github.com/openclaw/openclaw/releases';
+export const ADD_AGENT_TAB_ID = '__add_agent__';
+export const STATIC_DEFAULT_LOCAL_AGENT_HISTORY_INTEGRATIONS = ['openclaw'] as const;
+
