+++ packages/node-ui/src/ui/components/Shell/PanelRight/index.ts
@@ -1,0 +1,13 @@
+export { PanelRight } from './PanelRightContainer.js';
+export { ConnectedAgentsTab } from './ConnectedAgentsTab.js';
+export { NetworkTab } from './NetworkTab.js';
+export { SessionsTab } from './SessionsTab.js';
+export * from './attachments.js';
+export * from './chat-context.js';
+export * from './constants.js';
+export * from './format.js';
+export * from './messages.js';
+export * from './peer-model.js';
+export * from './selection.js';
+export * from './types.js';
+
